pyproject.toml
[tool.poetry]
name = "mnist-cnn"
version = "0.1.0"
description = "Predictions of mnist dataset using CNN"
authors = ["Parutin Maxim <parutin.ma@phystech.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
torch = "^2.0.1"
pandas = "^2.1.1"
torchvision = "^0.15.2"
urllib3 = "1.26.6"
fire = "^0.5.0"
hydra-core = "^1.3.2"
wandb = "^0.16.0"
torchmetrics = "^1.2.0"
scikit-learn = "^1.3.2"
mlflow = "^2.8.1"
pytorch-lightning = "^2.1.2"
pydantic = "1.10.11"
onnx = "^1.15.0"

[tool.poetry.dev-dependencies]
pre-commit = "^3.4.0"
dvc = "^3.30.3"
dvc-gdrive = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 75
target-version = ["py39"]

[tool.isort]
src_paths = ["ds_project"]
profile = "black"
line_length = 75
lines_after_imports = 2
